requirements.txt
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.1.2
django-allauth==0.42.0
django-crispy-forms==1.9.2
django-filter==2.4.0
django-seed==0.2.2
Faker==4.9.0
idna==2.10
isort==5.5.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
pylint==2.6.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.0
text-unidecode==1.3
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
wrapt==1.12.1
